requirements.txt
# ===============================================
# Crypto AI Trading Bot — Runtime Requirements
# Python 3.10+ compatible, minimal production deps
# ===============================================

# --- Core & Config ---
ccxt==4.4.98
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.10.1
pyyaml==6.0.2
orjson==3.10.12
requests==2.32.4
tenacity>=8.2.0,<9.0.0

# --- Redis / Streams / Realtime ---
redis==5.0.8
websockets>=13.0

# --- Async stack ---
aiohttp==3.12.15
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
async-timeout==4.0.3
attrs==25.3.0
frozenlist==1.7.0
multidict==6.6.3
yarl==1.20.1

# --- Numerical / Data ---
numpy==2.1.3
pandas==2.3.1
scipy==1.15.2

# --- TA & Modeling ---
ta-lib==0.5.1  # Technical analysis library (C library installed in Dockerfile, using newer compatible version)
ta==0.11.0
scikit-learn==1.7.1
scikit-optimize==0.10.2  # Bayesian optimization for parameter tuning
xgboost==3.0.3
joblib==1.5.1
statsmodels==0.14.5

# --- Observability / CLI ---
prometheus-client==0.22.1
rich==14.1.0
click==8.2.1
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2

# --- MCP / LLM hooks ---
mcp==1.12.4
openai==1.99.6
httpx==0.28.1
uvicorn==0.35.0
starlette==0.47.2
sse-starlette==3.0.2

# --- Agent Framework ---
autogen-agentchat==0.7.4
autogen-core==0.7.4
fastmcp==2.12.0
langgraph==0.6.6

# --- Sentiment / News ---
praw==7.8.1
vadersentiment==3.3.2
yfinance==0.2.65
fredapi==0.5.2

# --- Web3 ---
web3==7.13.0

# --- RAG (optional) ---
llama-index-core==0.12.0